projects/ml_backend/pyproject.toml
# Purpose: Declare the ML backend subproject package metadata and install surface.
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "birdsys-ml-backend"
version = "0.1.0"
description = "BirdSys served-model runtime"
requires-python = ">=3.11,<3.12"
dependencies = [
  "fastapi>=0.116.0,<1",
  "numpy>=1.26,<3",
  "opencv-python>=4.10,<5",
  "pillow>=10,<12",
  "pydantic>=2.10,<3",
  "python-dotenv>=1.0,<2",
  "pyyaml>=6.0,<7",
  "timm>=1.0.12,<2",
  "torch>=2.5,<2.11",
  "torchvision>=0.20,<0.26",
  "ultralytics>=8.3,<9",
  "uvicorn[standard]>=0.33,<1",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["birdsys.ml_backend*"]
namespaces = true
